move package do a different folder
