+++ packages/duckdb-core/src/duckdb-utils.ts
@@ -38,6 +38,62 @@ export function makeQualifiedTableName({
 }
 
 /**
+ * Returns the final identifier segment from a possibly-qualified SQL name.
+ *
+ * The parser is quote-aware: dots inside double-quoted identifiers are treated
+ * as part of the identifier rather than as qualification separators. Embedded
+ * quotes inside a quoted segment use the standard `""` escape.
+ *
+ * @example
+ * getUnqualifiedSqlIdentifier('schema.table')              // 'table'
+ * getUnqualifiedSqlIdentifier('db.schema.table')           // 'table'
+ * getUnqualifiedSqlIdentifier('schema."my.funny.table"')   // 'my.funny.table'
+ * getUnqualifiedSqlIdentifier('"weird""name"')             // 'weird"name'
+ */
+export function getUnqualifiedSqlIdentifier(
+  qualifiedName: string | undefined,
+): string | undefined {
+  if (!qualifiedName) return undefined;
+  const input = qualifiedName.trim();
+  if (!input) return undefined;
+
+  const parts: string[] = [];
+  let current = '';
+  let inQuotes = false;
+
+  for (let i = 0; i < input.length; i += 1) {
+    const ch = input[i];
+    if (ch === '"') {
+      // In SQL identifiers, escaped quotes inside quoted identifiers are doubled.
+      if (inQuotes && input[i + 1] === '"') {
+        current += '"';
+        i += 1;
+        continue;
+      }
+      inQuotes = !inQuotes;
+      current += ch;
+      continue;
+    }
+
+    if (ch === '.' && !inQuotes) {
+      parts.push(current);
+      current = '';
+      continue;
+    }
+
+    current += ch;
+  }
+  parts.push(current);
+
+  const last = parts[parts.length - 1]?.trim();
+  if (!last) return undefined;
+  if (last.startsWith('"') && last.endsWith('"') && last.length >= 2) {
+    return last.slice(1, -1).replaceAll('""', '"');
+  }
+  return last;
+}
+
+/**
  * Escapes a value for use in DuckDB SQL queries by wrapping it in single quotes
  * and escaping any existing single quotes by doubling them.
  *
