+++ packages/kepler/package.json
@@ -37,7 +37,6 @@
     "@kepler.gl/types": "3.3.0-alpha.0",
     "@kepler.gl/utils": "3.3.0-alpha.0",
     "@paralleldrive/cuid2": "^3.0.0",
-    "@sqlrooms/cells": "workspace:*",
     "@sqlrooms/duckdb-core": "workspace:*",
     "@sqlrooms/kepler-config": "workspace:*",
     "@sqlrooms/room-shell": "workspace:*",
